Dimaio/kinematics cpp (#191)
* Allow TView to wrap 2D eigen tensors

* linting

* Fix Tview logic when inputting 1x1 tensors.

* Remove unused variable

* enable_tensor_view now explicitly specifies the number of dimensions each view should squeeze

* Allow tensor views to explicitly specify consumed dimensions.  Use in both tpack construction and tview wrapping.

* Small reorganization in mgpu headers

* Missing include

* Add a generic segmented scan kernel

* linting

* Fix two bugs in segscan: 1) argument order of op was inconsistent, fine for add but no good for matmul; 2) bug when a segment boundaray aligned exactly with a CTA boundary.  Adding new tests to detect #2

* Add forward kinematics as an eigen/C++ kernel.  CUDA implementation uses segscan.

* linting

* linting(2)

* Update to jitted code; update to scan_ordering_test.

* linting

* Error in previous commit

* Update BackwardKin to eigen C++

* linting

* Moving derivative cart->dof mapping to C++

* linting

* Add derivative scanning pass

* Fixed file in previous commit

* Small bugfix with excl segmented scan with vt>1

* linting

* Clean up CUDA code to eliminate unnecessary lambdas

* Remove kinematics dependency on TCollection by passing a 1D concatentated vector to segscan routines

* Updates to scan ordering.  Small code cleanup in kin compiled headers.

* linting

* Small typo in original checkin

+++ tmol/kinematics/datatypes.py
@@ -53,12 +53,12 @@ class KinTree(TensorGroup, ConvertAttrs):
         frame_z = kin-atom index of grandparent
     """
 
-    id: Tensor(torch.long)[...]  # used as an index so long
+    id: Tensor(torch.int)[...]  # used as an index so long
     doftype: Tensor(torch.int)[...]
-    parent: Tensor(torch.long)[...]  # used as an index so long
-    frame_x: Tensor(torch.long)[...]
-    frame_y: Tensor(torch.long)[...]
-    frame_z: Tensor(torch.long)[...]
+    parent: Tensor(torch.int)[...]  # used as an index so long
+    frame_x: Tensor(torch.int)[...]
+    frame_y: Tensor(torch.int)[...]
+    frame_z: Tensor(torch.int)[...]
 
     @classmethod
     @convert_args
